+++ lib/generators/cypress_on_rails/templates/spec/cypress/app_commands/clean.rb
@@ -7,4 +7,6 @@ else
   Post.delete_all if defined?(Post)
 end
 
+CypressOnRails::SmartFactoryWrapper.reload
+
 Rails.logger.info "APPCLEANED" # used by log_fail.rb