+++ spec/integrations/rails_5_2/test.sh
@@ -11,7 +11,8 @@ cd $DIR
 
 echo '-- bundle install'
 bundle --version
-bundle install --quiet --gemfile="$DIR/Gemfile" --retry 2 --path vendor/bundle
+bundle config set --local path 'vendor/bundle'
+bundle install --quiet --gemfile="$DIR/Gemfile" --retry 2
 
 echo '-- migration'
 bundle exec ./bin/rails db:drop || true